Rename from `databricks_mws_budget` to `databricks_budget`
