+++ internal/providers/sdkv2/sdkv2.go
@@ -29,6 +29,7 @@ import (
 	"github.com/databricks/terraform-provider-databricks/commands"
 	"github.com/databricks/terraform-provider-databricks/common"
 	"github.com/databricks/terraform-provider-databricks/dashboards"
+	"github.com/databricks/terraform-provider-databricks/finops"
 	providercommon "github.com/databricks/terraform-provider-databricks/internal/providers/common"
 	"github.com/databricks/terraform-provider-databricks/jobs"
 	"github.com/databricks/terraform-provider-databricks/logger"
@@ -131,6 +132,7 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_azure_adls_gen1_mount":           storage.ResourceAzureAdlsGen1Mount().ToResource(),
 			"databricks_azure_adls_gen2_mount":           storage.ResourceAzureAdlsGen2Mount().ToResource(),
 			"databricks_azure_blob_mount":                storage.ResourceAzureBlobMount().ToResource(),
+			"databricks_budget":                          finops.ResourceBudget().ToResource(),
 			"databricks_catalog":                         catalog.ResourceCatalog().ToResource(),
 			"databricks_catalog_workspace_binding":       catalog.ResourceCatalogWorkspaceBinding().ToResource(),
 			"databricks_connection":                      catalog.ResourceConnection().ToResource(),
@@ -164,7 +166,6 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_mlflow_webhook":                  mlflow.ResourceMlflowWebhook().ToResource(),
 			"databricks_model_serving":                   serving.ResourceModelServing().ToResource(),
 			"databricks_mount":                           storage.ResourceMount().ToResource(),
-			"databricks_mws_budget":                      mws.ResourceMwsBudget().ToResource(),
 			"databricks_mws_customer_managed_keys":       mws.ResourceMwsCustomerManagedKeys().ToResource(),
 			"databricks_mws_credentials":                 mws.ResourceMwsCredentials().ToResource(),
 			"databricks_mws_log_delivery":                mws.ResourceMwsLogDelivery().ToResource(),
